feat: end of the challenge of activity 5 class 4
In that activity, the teacher challenged us to try to do the method to remove the book ourselfs. In this case, what I did was make a method very similar to the get by id, but remake the list of books, but filter out the book with the ID passed in the request param. In the end, the hteacher did exactly the same thing, so, maybe I did a good job? :thinking: